% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ecs.R
\name{stage.ecs_compute}
\alias{stage.ecs_compute}
\title{Title}
\usage{
\method{stage}{ecs_compute}(fn, cpu, memory,
  backend = default_compute_backend())
}
\arguments{
\item{backend}{}
}
\description{
Title
}
