% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_aws.R
\name{init_aws}
\alias{init_aws}
\title{Title}
\usage{
init_aws(storage_bucket, compute_cluster, compute_image,
  compute_execution_role, compute_subnets, compute_task_role,
  compute_assign_public_ip,
  region = aws.signature::locate_credentials()$region)
}
\arguments{
\item{compute_cluster}{}
}
\description{
Title
}
